add test EstimateGasForAction for payload

+++ gasstation/gasstattion_test.go
@@ -132,7 +132,7 @@ func TestEstimateGasForAction(t *testing.T) {
 	ret, err = gs.EstimateGasForAction(act)
 	require.NoError(err)
 	// base intrinsic gas 10000,plus data size*ExecutionDataGas
-	require.Equal(uint64(10000)+uint64(10*action.ExecutionDataGas), ret)
+	require.Equal(uint64(10000)+10*action.ExecutionDataGas, ret)
 }
 func getAction() (act *iotextypes.Action) {
 	pubKey1 := testaddress.Keyinfo["alfa"].PubKey
